Update zhttpmanager.cpp
The response size of some opcodes is more than 500KByte.
In this case, pushpin receives splited response stream from backend. (It is usually split into 8KByte size)
Because these individual packets do not have a normal JSON format, individual interpretation was not possible.
Therefore, in the past, these packets could not be cached and were transmitted to clients as they are.
To fix this, create a buffer for these splitted packets and keep pushing into the buffer until the full response is received.
Once the full response is received, it is reconstructed into one packet, then is interpreted as JSON format, and the entire response is stored in the cache.